fix: move components out of shad folder
